requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cached-property==2.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
frozendict==2.4.6
genanki==0.13.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.97.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
